+++ src/passes/DuplicateFunctionElimination.cpp
@@ -46,6 +46,7 @@ struct DuplicateFunctionElimination : public Pass {
     } else {
       limit = 1;
     }
+
     while (limit > 0) {
       limit--;
       // Hash all the functions